Add SimulatorError
