+++ crates/tests/Cargo.toml
@@ -14,6 +14,7 @@ veryl-formatter = {version = "0.19.0", path = "../formatter"}
 veryl-metadata  = {version = "0.19.0", path = "../metadata"}
 veryl-parser    = {version = "0.19.0", path = "../parser"}
 veryl-path      = {version = "0.19.0", path = "../path"}
+veryl-simulator = {version = "0.19.0", path = "../simulator"}
 veryl           = {version = "0.19.0", path = "../veryl"}
 
 [dev-dependencies]
