+++ testcases/error/recursive_function.veryl
@@ -1,0 +1,12 @@
+module recursive_function (
+    a: input  logic<32>,
+    c: output logic<32>,
+) {
+    function recurse(x: input logic<32>) -> logic<32> {
+        return recurse(x);
+    }
+
+    always_comb {
+        c = recurse(a);
+    }
+}
