+++ testcases/error/top_module_not_found.veryl
@@ -1,0 +1,6 @@
+module SomeOtherName (
+    a: input  logic<8>,
+    b: output logic<8>,
+) {
+    assign b = a + 1;
+}
